Make tool change, configuration, and heater management more solid:
- Temperature module is initialized in full or with no heater based on legacy head ID
- Temperatue ISR is correctly enabled when enablig heaters/sensors
- Min & max temp in addon definiton structures are made both signed int
- Default value for no measured temperature is a more user-friendly '0'

+++ Marlin/tools.h
@@ -18,8 +18,8 @@ typedef struct tool_s {
    uint8_t extruders:4;
    uint8_t heaters:4;
    uint8_t thtable = THERMISTOR_HOTSWAP_DEFAULT_INDEX;
-   int8_t  mintemp = HEATER_0_MINTEMP;
-   int16_t maxtemp = HEATER_0_MAXTEMP;
+   int16_t mintemp = HEATER_0_MINTEMP;  // °C
+   int16_t maxtemp = HEATER_0_MAXTEMP;  // °C
    const char *mods = NULL;
 } tool_t;
